np.int -> np.int32

+++ mmpose/datasets/pipelines/bottom_up_transform.py
@@ -151,7 +151,7 @@ class HeatmapGenerator:
             self.output_size = output_size
         else:
             self.output_size = np.array([output_size, output_size],
-                                        dtype=np.int)
+                                        dtype=np.int32)
         self.num_joints = num_joints
         if sigma < 0:
             sigma = self.output_size.prod()**0.5 / 64
@@ -231,7 +231,7 @@ class JointsEncoder:
             self.output_size = output_size
         else:
             self.output_size = np.array([output_size, output_size],
-                                        dtype=np.int)
+                                        dtype=np.int32)
         self.tag_per_joint = tag_per_joint
 
     def __call__(self, joints):
@@ -282,7 +282,7 @@ class PAFGenerator:
             self.output_size = output_size
         else:
             self.output_size = np.array([output_size, output_size],
-                                        dtype=np.int)
+                                        dtype=np.int32)
         self.limb_width = limb_width
         self.skeleton = skeleton
 
@@ -388,7 +388,7 @@ class BottomUpRandomFlip:
                     assert len(_output_size) == 2
                 else:
                     _output_size = np.array([_output_size, _output_size],
-                                            dtype=np.int)
+                                            dtype=np.int32)
                 mask[i] = mask[i][:, ::-1].copy()
                 joints[i] = joints[i][:, self.flip_index]
                 joints[i][:, :, 0] = _output_size[0] - joints[i][:, :, 0] - 1
@@ -730,7 +730,7 @@ class BottomUpGetImgSize:
         if input_size.size > 1:
             assert len(input_size) == 2
         else:
-            input_size = np.array([input_size, input_size], dtype=np.int)
+            input_size = np.array([input_size, input_size], dtype=np.int32)
         img = results['img']
 
         h, w, _ = img.shape
@@ -798,7 +798,7 @@ class BottomUpResizeAlign:
         if input_size.size > 1:
             assert len(input_size) == 2
         else:
-            input_size = np.array([input_size, input_size], dtype=np.int)
+            input_size = np.array([input_size, input_size], dtype=np.int32)
         test_scale_factor = results['ann_info']['test_scale_factor']
         aug_data = []
